% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACSdata-package.R
\docType{package}
\name{MACSdata}
\alias{MACSdata}
\title{MACSdata}
\description{
This is an ExperimentHub Package for files that used in MACS unit
test. Details can be found at: https://macs3-project.github.io/macs/.
}
